pyproject.toml
[tool.poetry]
name = "hotels-admin-project"
version = "0.1.0"
description = ""
authors = ["Bullochka <tamer.kab@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
